Fix gossip network event overwriting self
When a node joins a network it sends out a gossip event before it
updates it's own in memory state. This can create race where the node
gets the event back from a remote node before we update in memory state
and we treat that as latest state. To avoid this race, always generate
the gossip after updating local state.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>